control-plane-app/requirements.txt
# Python dependencies for AI Control Plane App
# Install with: pip install -r requirements.txt

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
python-multipart>=0.0.6
websockets>=12.0
httpx>=0.27.0
databricks-sdk>=0.40.0
